Semantic-Hotel-Search-main/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
